requirements.txt
python-dotenv
nltk
spacy
requests
unidecode
lxml
numpy
pandas
xlrd
openpyxl
tqdm
beautifulsoup4
plotly
matplotlib
kaleido